Fix the way the start address for the given network is parsed
The CURR_STARTADDR needs to be an address without the network suffix.
It was being set to the user-provided value, which could now include a slash (/).
With this change, the slash and anything after it are safely removed from that
variable before it is used.

This was causing the OLSRd configuration to be invalid, making the deamon not to
start at all.